mount.cifs: clean up command-line options
The mount.cifs command apparently tries to take a ton of command-line
options. Many of these will never be passed to mount.cifs by /bin/mount.
Others are more appropriately specified as mount options.

In both cases, there are a lot of options in the switch statement that
are not listed in the optstring, and there are characters in the
optstring that are not dealt with by the switch statement. Other options
are poorly wired to the rest of the code and don't actually do anything.

Clean it up by removing all but the ones that are likely to ever be
used.

Signed-off-by: Jeff Layton <jlayton@redhat.com>