cifs-utils: add a note about inclusion of keys.dns_resolver program in keyutils
As of version 1.5, the keyutils package is shipping a generic
dns_resolver upcall. Add a note to the cifs.upcall manpage that mentions
this and recommends the use of that program over cifs.upcall.

Eventually, we may want to be able to conditionally compile out the
dns_resolver part of the upcall, but it's already pretty small and
wouldn't save us very much.

Signed-off-by: Jeff Layton <jlayton@samba.org>