cifsacl: Add configure and make directives for cifsacl (try #2)
Add configure directives for option cifsacl.  The default action is
to enable cifsacl option.
cifsacl option is enabled or disabled in a similar way to cifs.idmap
in the same function. In addition, for cifsacl, check for sys/xattr.h
is done in the smae .m4 file.

Add directives to build getcifsacl in Makefile.

Signed-off-by: Shirish Pargaonkar <shirishpargaonkar@gmail.com>