autoconf: work around broken wbclient.h file
Some versions of wbclient.h have function declarations with bool type
args, but they don't include stdbool.h themselves. Make sure that
we can deal with that by telling the autoconf test to include stdbool.h
explicitly. In order to do that properly we need to move some of the
standard header and type tests up in the file.

Signed-off-by: Jeff Layton <jlayton@samba.org>