cifs.upcall: move to Andrew's suggested algorithm for picking a principal
Andrew Bartlett suggests the heuristic supplied in the comments.

For now, we don't try to guess the domainname when the hostname is not
qualified, but add a comment with what needs to be done in order to
support that.

Also, with this change we no longer need util.o to be linked in.

Signed-off-by: Jeff Layton <jlayton@samba.org>