cifs.upcall: try and guess the domain name on unqualified names
Resolve the unqualified hostname and set AI_CANONNAME to make sure that
field is populated. Scan forward to the first '.' in ai_canonname, and
append that value onto the unqualified hostname to get a FQDN. Then
prepend that value with "cifs/" and try to get a service ticket for
that principal.

Signed-off-by: Jeff Layton <jlayton@samba.org>