manpage: clarify usage of the username= option
Specifying the username as "user=" can confuse some versions of /bin/mount,
and cause it to append noexec to the mount string. Also, remove the blurb
about how cifs.ko will accept abbreviations -- it's just bad practice since
it leads to these sorts of conflicts.

A wonderful example of this confusion was reported by Stefan here:

    https://bugzilla.redhat.com/show_bug.cgi?id=860154

Reported-by: Stefan Walter <walteste@inf.ethz.ch>
Signed-off-by: Jeff Layton <jlayton@samba.org>